+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
 
+## [0.3.2] - 2026-02-12
+### Fixed
+- Fix Cell disassembly overwriting the whole stack to empty, instead of just one cell.
+
+
 ## [0.3.1] - 2026-02-11
 ### Added
 - Add Import Interface texture and fix the recipe.